move media upload widget to its own file
