+++ frontend/app/features/archive/pages/CreateProductionPage.tsx
@@ -20,12 +20,7 @@ import EventSection from "../components/EventSection";
 import { ProductionGeneralInfo } from "../components/ProductionGeneralInfo";
 import { ARCHIVE_PERMISSIONS } from "../archive.constants";
 import { uploadMedia } from "../services/mediaService";
-import PermMediaOutlinedIcon from "@mui/icons-material/PermMediaOutlined";
-import CloseIcon from "@mui/icons-material/Close";
-import StarBorderIcon from "@mui/icons-material/StarBorder";
-import StarIcon from "@mui/icons-material/Star";
-import IconButton from "@mui/material/IconButton";
-import Tooltip from "@mui/material/Tooltip";
+import { MediaUploadWidget } from "../components/MediaUploadWidget";
 
 function isInfoModified(draftInfo: ProductionInfo | null): boolean {
   if (!draftInfo) return false;
@@ -38,189 +33,6 @@ function isInfoModified(draftInfo: ProductionInfo | null): boolean {
     !!draftInfo.teaser ||
     !!draftInfo.description ||
     !!draftInfo.info
-  );
-}
-
-// ─── Media Upload Widget ───────────────────────────────────────────────────
-
-type MediaPreview = { src: string; isVideo: boolean; file: File };
-
-type MediaUploadWidgetProps = {
-  onFilesChange: (files: File[]) => void;
-  onBannerUrlChange: (url: string | undefined) => void;
-};
-
-/**
- * Files are held in browser memory only. `onFilesChange` is called with the
- * full ordered list every time it changes. The banner (first file) is tracked
- * with a star toggle — selecting a new banner moves that file to index 0 while
- * keeping the relative order of the rest.
- */
-function MediaUploadWidget({
-  onFilesChange,
-  onBannerUrlChange,
-}: MediaUploadWidgetProps) {
-  const { t } = useTranslation();
-  const fileInputRef = useRef<HTMLInputElement>(null);
-  const [previews, setPreviews] = useState<MediaPreview[]>([]);
-  const [isDragging, setIsDragging] = useState(false);
-
-  function notify(updated: MediaPreview[]) {
-    onFilesChange(updated.map((p) => p.file));
-    onBannerUrlChange(updated[0]?.isVideo ? undefined : updated[0]?.src);
-  }
-
-  function handleFiles(files: FileList | null) {
-    if (!files) return;
-    const incoming = Array.from(files)
-      .filter((f) => f.type.startsWith("image/") || f.type.startsWith("video/"))
-      .map((f) => ({
-        src: URL.createObjectURL(f),
-        isVideo: f.type.startsWith("video/"),
-        file: f,
-      }));
-    setPreviews((prev) => {
-      const updated = [...prev, ...incoming];
-      notify(updated);
-      return updated;
-    });
-  }
-
-  function removePreview(index: number) {
-    setPreviews((prev) => {
-      URL.revokeObjectURL(prev[index].src);
-      const updated = prev.filter((_, i) => i !== index);
-      notify(updated);
-      return updated;
-    });
-  }
-
-  /** Move the chosen item to position 0 (banner slot). */
-  function setBanner(index: number) {
-    if (index === 0) return;
-    setPreviews((prev) => {
-      const updated = [prev[index], ...prev.filter((_, i) => i !== index)];
-      notify(updated);
-      return updated;
-    });
-  }
-
-  return (
-    <div className="mt-2">
-      <div className="mb-2 flex items-center justify-between">
-        <h3 className="text-archive-ink/70 text-xs font-bold tracking-[0.2em] uppercase">
-          {t("archive.media.title")}
-        </h3>
-        {previews.length > 0 && (
-          <p className="text-archive-ink/40 text-xs">{t("archive.media.bannerHint")}</p>
-        )}
-      </div>
-
-      <div
-        onDragOver={(e) => {
-          e.preventDefault();
-          setIsDragging(true);
-        }}
-        onDragLeave={() => setIsDragging(false)}
-        onDrop={(e) => {
-          e.preventDefault();
-          setIsDragging(false);
-          handleFiles(e.dataTransfer.files);
-        }}
-        onClick={() => fileInputRef.current?.click()}
-        className={`flex cursor-pointer flex-col items-center justify-center gap-2 rounded-2xl border-2 border-dashed px-6 py-8 transition ${
-          isDragging
-            ? "border-archive-accent bg-archive-accent/10"
-            : "border-archive-ink/15 hover:border-archive-accent/60 bg-archive-ink/3"
-        }`}
-      >
-        <PermMediaOutlinedIcon
-          className="text-archive-ink/40"
-          style={{ fontSize: 36 }}
-        />
-        <p className="text-archive-ink/60 text-center text-sm leading-snug">
-          {t("archive.media.dropMedia")}{" "}
-          <span className="text-archive-accent font-semibold underline">
-            {t("archive.media.browse")}
-          </span>
-        </p>
-        <p className="text-archive-ink/35 text-xs">
-          {t("archive.media.acceptedFormats")}
-        </p>
-        <input
-          ref={fileInputRef}
-          type="file"
-          accept="image/*,video/*"
-          multiple
-          className="hidden"
-          onChange={(e) => handleFiles(e.target.files)}
-        />
-      </div>
-
-      {previews.length > 0 && (
-        <div className="mt-3 grid grid-cols-3 gap-2 sm:grid-cols-4 md:grid-cols-5">
-          {previews.map((item, i) => (
-            <div
-              key={item.src}
-              className={`relative aspect-square overflow-hidden rounded-xl bg-black ${
-                i === 0 ? "ring-archive-accent ring-2" : ""
-              }`}
-            >
-              {item.isVideo ? (
-                <video
-                  src={item.src}
-                  className="h-full w-full object-cover"
-                  muted
-                  preload="metadata"
-                />
-              ) : (
-                <img
-                  src={item.src}
-                  alt=""
-                  className="h-full w-full object-cover transition hover:scale-105"
-                />
-              )}
-
-              {/* Banner toggle */}
-              <Tooltip
-                title={
-                  i === 0 ? t("archive.media.isBanner") : t("archive.media.setBanner")
-                }
-              >
-                <IconButton
-                  size="small"
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    setBanner(i);
-                  }}
-                  className="!absolute !bottom-1 !left-1 !bg-black/50 !text-yellow-400 hover:!bg-black/70"
-                >
-                  {i === 0 ? (
-                    <StarIcon style={{ fontSize: 14 }} />
-                  ) : (
-                    <StarBorderIcon style={{ fontSize: 14 }} />
-                  )}
-                </IconButton>
-              </Tooltip>
-
-              {/* Remove */}
-              <Tooltip title={t("archive.media.remove")}>
-                <IconButton
-                  size="small"
-                  onClick={(e) => {
-                    e.stopPropagation();
-                    removePreview(i);
-                  }}
-                  className="!absolute !top-1 !right-1 !bg-black/50 !text-white hover:!bg-black/70"
-                >
-                  <CloseIcon style={{ fontSize: 14 }} />
-                </IconButton>
-              </Tooltip>
-            </div>
-          ))}
-        </div>
-      )}
-    </div>
   );
 }
 
@@ -290,7 +102,8 @@ async function handleAddProduction(
     // Go to newly created production
     const currentParts = window.location.pathname.split("/");
     const responseParts = response["id_url"].split("/");
-    currentParts[currentParts.length - 1] = responseParts[responseParts.length - 1];
+    currentParts[currentParts.length - 1] =
+      responseParts[responseParts.length - 1];
     navigate(currentParts.join("/"));
   } catch (e) {
     window.alert(`${errorMessage}: ${e}`);
@@ -344,9 +157,7 @@ export function CreateProductionPage() {
   const [draftTags, setDraftTags] = useState<Tag[]>([]);
   const [draftEvents, setDraftEvents] = useState<EventWithResolvedRelations[]>([]);
   const [mediaFiles, setMediaFiles] = useState<File[]>([]);
-  const [bannerPreviewUrl, setBannerPreviewUrl] = useState<string | undefined>(
-    undefined
-  );
+  const [bannerPreviewUrl, setBannerPreviewUrl] = useState<string | undefined>(undefined);
 
   const [isQuillDirty, setIsQuillDirty] = useState(false);
   const [isSaving, setIsSaving] = useState<boolean>(false);
@@ -417,10 +228,7 @@ export function CreateProductionPage() {
           {t("archive.create_production")}
         </h1>
         <ProductionHeader
-          image_url={
-            bannerPreviewUrl ??
-            "data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7"
-          }
+          image_url={bannerPreviewUrl ?? "data:image/gif;base64,R0lGODlhAQABAIAAAAAAAP///yH5BAEAAAAALAAAAAABAAEAAAIBRAA7"}
           isEditing={true}
           originalInfo={null}
           draftInfo={draftInfo}
